change the data type of frequentRenterPoints to doule from int.
add the abstract method getFrequentRenterPointsStep in MovieStrategy instead of the default step 1 in BaseStatement

+++ src/main/java/rentalstore/statement/HtmlStatement.java
@@ -21,7 +21,7 @@ public class HtmlStatement extends BaseStatement {
     }
 
     @Override
-    public String getFooterString(double totalAmount, int frequentRenterPoints) {
+    public String getFooterString(double totalAmount, double frequentRenterPoints) {
         return "<P>You owe<EM>" + String.valueOf(totalAmount) + "</EM><P>\n"
                 + "On this rental You earned<EM>" + String.valueOf(frequentRenterPoints) + "</EM> frequent renter points<P>";
     }
